requirements.txt
# --- Core OCR and Imaging ---
pillow==10.3.0
opencv-python==4.10.0.84
mss==9.0.1
dxcam==0.0.5 ; platform_system == "Windows"
numpy==1.26.4

# --- GUI ---
PySide6==6.7.2

# --- Translation ---
googletrans==4.0.0-rc1

# --- Math Formula Support ---
latex2mathml==3.77.0

# --- Utilities ---
python-dotenv==1.0.1
packaging==24.1
pywin32==306 ; platform_system == "Windows"

# --- Optional Hindi Postprocessing ---
indic-nlp-library==0.92
pypandoc
